# backend-jee/openapi.yaml
openapi: 3.0.3
info:
  title: Test Management API
  description: |
    A comprehensive test management system for creating, managing, and administering tests.
    
    ## Features
    - Enterprise and domain management
    - Test creation and management
    - Question management (QCM, True/False, Open questions)
    - Candidate invitation system
    - Test session management
    - AI-powered scoring with Groq
    - Detailed test results and analytics
    
    ## Authentication
    Most endpoints require authentication via cookie-based JWT tokens.
    - Admin endpoints: Require admin authentication
    - Candidate endpoints: Require candidate authentication

  version: 1.0.0
  contact:
    name: API Support
    email: support@testmanagement.com
servers:
  - url: http://localhost:8080/api-rest-1.0-SNAPSHOT
    description: Local development server
  - url: https://api.testmanagement.com/api-rest-1.0-SNAPSHOT
    description: Production server
...
security:
  - cookieAuth: []
paths:
...
  /auth/test/{testId}/questions:
    get:
      tags:
        - Authentication
      summary: Get test questions (candidate view)
      description: Get all questions for a test without showing correct answers
      operationId: getTestQuestionsForCandidate
      security: []
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: integer
          description: Test ID
      responses:
        '200':
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CandidateQuestionResponse'

  # ============================================================
  # ENTERPRISE ENDPOINTS
  # ============================================================
...
  /test:
    post:
      tags:
        - Test Management
      summary: Create test
      description: Create a new test with questions
      operationId: createTest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestRequest'
      responses:
        '201':
          description: Test created successfully
          content:
            application/json:
              schema:
                type: string
                example: "Test created successfully with id 1"

    get:
      tags:
        - Test Management
      summary: Get all tests
      description: Get all tests for the authenticated enterprise
      operationId: getAllTests
      responses:
        '200':
          description: List of tests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestResponse'
  /test/{testId}/questions:
    get:
      tags:
        - Test Management
      summary: Get questions by test ID
      description: Get all questions in a specific test (admin view with correct answers)
      operationId: getQuestionsByTestId
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: integer
          description: Test ID
      responses:
        '200':
          description: List of questions with answers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestQuestionResponse'
  /test/{testId}/sessions:
    get:
      tags:
        - Test Management
      summary: Get test sessions
      description: Get all test sessions for a specific test with candidate details
      operationId: getTestSessions
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: integer
          description: Test ID
      responses:
        '200':
          description: List of test sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestSessionResponse'
  /test/invitation:
    post:
      tags:
        - Test Management
      summary: Send test invitations
      description: Invite candidates to take a test via email
      operationId: sendInvitations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestSessionRequest'
      responses:
        '201':
          description: Invitations sent successfully
          content:
            application/json:
              schema:
                type: string
                example: "Test invitation sent successfully"

  # ============================================================
  # QUESTION MANAGEMENT ENDPOINTS
  # ============================================================
  /test/questions:
    post:
      tags:
        - Question Management
      summary: Create question
      description: Create a new question (QCM, True/False, or Open)
      operationId: createQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionCreateRequest'
            examples:
              qcm:
                summary: Multiple Choice Question
                value:
                  label: "What is polymorphism?"
                  hint: "Think about OOP"
                  points: 10.0
                  questionType: "QCM"
                  answers:
                    - label: "Ability to take many forms"
                      correct: true
                    - label: "A loop"
                      correct: false
              truefalse:
                summary: True/False Question
                value:
                  label: "Java is platform-independent"
                  hint: "Think about JVM"
                  points: 5.0
                  questionType: "TRUE_OR_FALSE"
                  answers:
                    - label: "True"
                      correct: true
                    - label: "False"
                      correct: false
              open:
                summary: Open Question
                value:
                  label: "Explain the JVM"
                  hint: "Think about execution"
                  points: 10.0
                  questionType: "OPEN_QUESTION"
                  openAnswers:
                    expectedAnswer: "The JVM is a runtime environment..."
                    keyWords:
                      - keyword: "bytecode"
                      - keyword: "platform independence"
      responses:
        '201':
          description: Question created successfully
          content:
            application/json:
              schema:
                type: string

    get:
      tags:
        - Question Management
      summary: Get all questions
      description: Get all questions for the authenticated enterprise
      operationId: getAllQuestions
      responses:
        '200':
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuestionResponse'

  # ============================================================
  # CANDIDATE ENDPOINTS
  # ============================================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: test_token
      description: JWT token stored in cookie
  schemas:
...
    TestRequest:
      type: object
      required:
        - name
        - duration_minute
      properties:
        name:
          type: string
          example: "Java Programming Fundamentals"
        duration_minute:
          type: integer
          example: 60
        questions:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              position:
                type: integer
    TestResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        durationMinute:
          type: integer
        isActive:
          type: boolean
        isPublic:
          type: boolean
        enterpriseId:
          type: string
          format: uuid
        adminAccountId:
          type: integer
          format: int64
    TestSessionRequest:
      type: object
      required:
        - testId
        - emailCandidate
        - dateExpiration
      properties:
        testId:
          type: integer
          example: 1
        emailCandidate:
          type: array
          items:
            type: string
            format: email
          example: ["candidate1@email.com", "candidate2@email.com"]
        dateExpiration:
          type: string
          format: date-time
          example: "2026-02-01T23:59:59"
    TestSessionResponse:
      type: object
      properties:
        sessionId:
          type: integer
          format: int64
        candidateEmail:
          type: string
          format: email
        accessCode:
          type: string
        startTime:
          type: string
          format: date-time
        expirationTime:
          type: string
          format: date-time
        isUsed:
          type: boolean
        status:
          type: string
          enum: [PLANNED, SCHEDULED, ACTIVE, FINISHED]
        score:
          type: number
          format: double
        testName:
          type: string
        testId:
          type: integer
        testDuration:
          type: integer

    # ==================== Question Schemas ====================
    QuestionCreateRequest:
      type: object
      required:
        - label
        - points
        - questionType
      properties:
        label:
          type: string
          example: "What is polymorphism?"
        hint:
          type: string
          example: "Think about OOP"
        points:
          type: number
          format: double
          example: 10.0
        questionType:
          type: string
          enum: [QCM, TRUE_OR_FALSE, OPEN_QUESTION]
        answers:
          type: array
          items:
            $ref: '#/components/schemas/AnswerCreateRequest'
        openAnswers:
          $ref: '#/components/schemas/OpenAnswerCreateRequest'
    AnswerCreateRequest:
      type: object
      required:
        - label
        - correct
      properties:
        label:
          type: string
          example: "Ability to take many forms"
        correct:
          type: boolean
          example: true
    OpenAnswerCreateRequest:
      type: object
      properties:
        expectedAnswer:
          type: string
        keyWords:
          type: array
          items:
            type: object
            properties:
              keyword:
                type: string
    QuestionResponse:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        hint:
          type: string
        types:
          type: string
          example: "qcm"
        points:
          type: number
          format: double
        answers:
          type: array
          items:
            $ref: '#/components/schemas/AnswerResponse'
        openAnswers:
          $ref: '#/components/schemas/OpenAnswerResponse'
    TestQuestionResponse:
      type: object
      properties:
        position:
          type: integer
        id:
          type: integer
        label:
          type: string
        hint:
          type: string
        types:
          type: string
        points:
          type: number
        answers:
          type: array
          items:
            $ref: '#/components/schemas/AnswerResponse'
    CandidateQuestionResponse:
      type: object
      description: Question response for candidates (no correct answers shown)
      properties:
        id:
          type: integer
        label:
          type: string
        hint:
          type: string
        questionType:
          type: string
        points:
          type: number
        answers:
          type: array
          items:
            $ref: '#/components/schemas/CandidateAnswerResponse'
    AnswerResponse:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        correct:
          type: boolean
    CandidateAnswerResponse:
      type: object
      description: Answer response for candidates (no correct flag)
      properties:
        id:
          type: integer
        label:
          type: string
    OpenAnswerResponse:
      type: object
      properties:
        id:
          type: integer
        expectedAnswer:
          type: string
        keyWords:
          type: array
          items:
            $ref: '#/components/schemas/KeywordResponse'
    KeywordResponse:
      type: object
      properties:
        id:
          type: integer
        keyword:
          type: string

    # ==================== Candidate Test Submission ====================
...